+++ src/media/data/media/contents/genericv/crew.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 
 #
-# Copyright 2022 Chris Josephes
+# Copyright 2023 Chris Josephes
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
@@ -31,6 +31,7 @@ of visual art, like a movie or television show.
 # pylint: disable=R0801
 # pylint: disable=too-many-instance-attributes
 
+import xml.etree.ElementTree as ET
 from media.xml.functions import xs_bool
 from media.xml.namespaces import Namespaces
 from media.data.nouns import Name
@@ -160,6 +161,10 @@ class Role():
                         self.portrays.append(PortraysNarrator())
                     elif child_tag == 'self':
                         self.portrays.append(PortraysSelf(child))
+                    elif child_tag == 'background':
+                        self.portrays.append(PortraysBackground())
+                    elif child_tag == 'additionalVoices':
+                        self.portrays.append(PortraysAdditionalVoices())
                     elif child_tag == 'title':
                         self.portrays.append(CharacterTitle(child))
                     else:
@@ -211,6 +216,40 @@ class PortraysNarrator(Portrays):
         self.value = 'Narrator'
         self.formal = 'Narrator'
         self.sort_value = 'narrator'
+
+    def to_element(self):
+        '''
+        Generate an empty additionalVoices element.
+        '''
+        return ET.Element('additionalVoices')
+
+
+class PortraysBackground(Portrays):
+    '''
+    Class for special background character role.
+    '''
+    def __init__(self):
+        super().__init__()
+        self.value = 'Background'
+        self.formal = 'Background'
+        self.sort_value = 'background'
+
+    def to_element(self):
+        '''
+        Generate an empty background element.
+        '''
+        return ET.Element('background')
+
+
+class PortraysAdditionalVoices(Portrays):
+    '''
+    Class for special additionalVoices character role.
+    '''
+    def __init__(self):
+        super().__init__()
+        self.value = 'Additional Voices'
+        self.formal = 'Additional Voices'
+        self.sort_value = 'additional voices'
 
 
 class PortraysSelf(Portrays):